# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SWAN API
  version: 1.0.0
  description: Auto-generated API documentation
paths:
...
  /v1/studies/{id}/{index}/:
    get:
      operationId: v1_studies_retrieve_entry
      description: The file at the index of the dataset based on the user
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: A UUID string identifying this study.
        required: true
      - in: path
        name: index
        schema:
          type: string
          pattern: ^[0-9]+$
        required: true
      tags:
      - Study
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
          description: ''
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid